+++ web-client/trunk/src/main/java/com/cityopt/service/TypeServiceImpl.java
@@ -25,8 +25,23 @@ public class TypeServiceImpl implements TypeService {
 	}
 
 	@Transactional
-	public void delete(Type u) {
+	public void delete(Type u) throws EntityNotFoundException {
+		
+		if(typeRepository.findOne(u.getTypeid()) == null) {
+			throw new EntityNotFoundException();
+		}
+		
 		typeRepository.delete(u);
+	}
+	
+	@Transactional
+	public Type update(Type toUpdate) throws EntityNotFoundException {
+		
+		if(typeRepository.findOne(toUpdate.getTypeid()) == null) {
+			throw new EntityNotFoundException();
+		}
+		
+		return save(toUpdate);
 	}
 	
 	public Type findByID(Integer id) {
